+++ packages/yourloops/pages/notifications/notification.tsx
@@ -40,7 +40,7 @@ import IconButton from "@material-ui/core/IconButton";
 import Button from "@material-ui/core/Button";
 import Tooltip from "@material-ui/core/Tooltip";
 
-import { IUser, UserRoles } from "../../models/shoreline";
+import { IUser, UserRoles } from "../../models/user";
 import { INotification, NotificationType } from "../../lib/notifications/models";
 import { errorTextFromException, getUserFirstName, getUserLastName } from "../../lib/utils";
 import { useNotification } from "../../lib/notifications/hook";