+++ packages/yourloops/test/pages/profile/profile.test.tsx
@@ -34,7 +34,7 @@ import { BrowserRouter } from "react-router-dom";
 import { Units } from "../../../models/generic";
 import { loggedInUsers } from "../../common";
 import ProfilePage from "../../../pages/profile";
-import { Profile, UserRoles } from "../../../models/shoreline";
+import { Profile, UserRoles } from "../../../models/user";
 import * as authHookMock from "../../../lib/auth";
 import User from "../../../lib/auth/user";
 import { HcpProfession } from "../../../models/hcp-profession";